contextual-embeddings-agentic-workflow/contextual_pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contextual-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.7.0",
    "pydantic-ai[anthropic]>=0.0.14",
    "neo4j>=5.20.0",
    "openai>=1.30.0",
    "tenacity>=8.3.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
